Fix keyerror on byteOffset
When importing the samples, there is now a crash due to a JSON keyerror.
This is because in the final glTF 2.0 specification, the byteOffset key
is optional on the first buffer view (when the offset is 0.)

This fix prevents the crash by first checking to make sure the key
exists, and if not just sets it to 0.