Scene/node forest improvements
Make a single armature object to hold the single glTF node forest.
Scenes are distinguished by only linking in the objects in a tree
they contain. Otherwise, every scene now gets a whole copy of the
armature.

Should solve the problem of two scenes which share a tree of the
node forest and the problem of one object being skinned by
nodes from a different tree than the one containing it.