Use one bmesh through all mesh creation
Instead of creating a mesh for each primitive, have one bmesh and
add all the data for each primitive into that one bmesh. This
requires relabelling indices but after #29 we already have to do
that so it's no big deal. This lets us concentrate the hack for
creating shape keys in one place and is also slightly faster, at
least for big models like MetalRoughSpheres, presumably because
of less round-tripping.