Work on correct world space positions
In particular, meshes and cameras that are children of bones should
be in the correct place now. My understanding of how bone rotation
worked advanced, as addressed in a long comment, so we now only allow
certain kinds of rotation and they interact correctly with scaling
in animation creation.

Vforest creation was factored into a seperate file since it has become
the most complicated part of the importer.