Better armature locations
Since I couldn't get confirmation that the skeleton node for a skin
is an ancestor of the joints when extant, we no longer depend on this.
We now use the common ancestor of all the joints and the skeleton as
the node where the arma should go. This doesn't change anything if
the skleton really is an ancestor.

If possible we also now insert the armature _below_ the selected
node instead of above.

I also found and fixed a place that required the skeleton property
to be present.