Node group serialization improvements
Taught the serializer about Mapping nodes which will be used in
the next commit. The process of editing the groups was made
simpler (see the comment). node_groups.json will be more line-
based from now on too; one line per group. This makes diffs
on it a little nicer.