Shore up KHR_texture_transform
Texture transforms are now tested. In particular a really stupid
bug was fixed that made us look for the extension on the wrong
object! Apparently no one else tested that it worked either :p

The texture transform nodes were moved into a frame and the
glTF<->Blender UV node was put into a group so the node tree
is a bit prettier.