Rewrite bone stuff
Heterogeneous scalings are handled by putting the model in a non-
rest pose in Blender that matches the rest pose in the glTF. Models
without heterogeneous scalings should still match the rest pose.
Also constrained skinned meshes to their armatures so it shouldn't
matter where in the tree they are instantiated.