Work on automatic bone rotation
Remove the manual selection of bone direction, which is complicated
and nobody uses. Move the calculation of how to interchange scales
and rotations to an explicit helper function. And try to pick better
targets to rotate towards. Also, bugfixes to the AXIS_TO_PLUS_Y
map, which had some bad values.