Add animations to NLA tracks
This aids discoverability for both humans and export scripts, which
often look there to find the actions that an object should use.

The only problem is "disposal". After one action plays, if the
subsequent does not animate (say) a bone, that bone stays in the
pose it finished the previous animation with. This means that
scrubbing backwards can give different results than scrubbing
forward.

Maybe I'll ask about this on the stack exchange....